requirements.txt

gunicorn==19.9.0
amqp==5.0.6
asgiref==3.4.1
attrs==21.4.0
billiard==3.6.4.0
boto3==1.17.52
botocore==1.20.52
celery==5.1.2
certifi==2019.6.16
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.4.2
Django==3.2.6
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
gunicorn==20.1.0
psycopg2-binary==2.9.1
requests==2.22.0
whitenoise==3.3.1
pytest-django==4.5.2





